Implement handling of instance usages in method refactoring now with a CHECK
This is not finished yet, there are 2 tests failing but at least no errors anymore

In this commit, we introduced the fHasInstanceUsages flag to track if a method being refactored has any instance variables or methods. If fHasInstanceUsages is true, the method will be modified to accept an additional parameter representing the instance of the class.

Key changes include:

Implementing an ASTVisitor to visit SimpleName nodes within the method body and check for instance variables or methods
If an instance variable is found, the code replaces the variable with an appropriate FieldAccess expression
If an instance method is found, the code replaces the method invocation with an appropriate MethodInvocation expression

This update improves the refactoring process by correctly handling methods with instance usages, enabling a more accurate static method conversion.