InititalConditionsChecker is now instantiated
all checks in checkInitialConditions are now more concise and easier to read. The checker methods are instance methods now and return a boolean value instead of the RefactoringStatus. The InititalConditionsChecker is instantiated with a RefactoringStatus object and uses it in every checker method to determine which boolean value needs to be returned.